requirements.txt
django<5.0
google-generativeai
